+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/tasks/SentryUploadSnapshotsTask.kt
@@ -10,6 +10,7 @@ import org.gradle.api.provider.Property
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.Optional
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity
 import org.gradle.api.tasks.TaskProvider
@@ -29,11 +30,30 @@ abstract class SentryUploadSnapshotsTask : SentryCliExecTask() {
   @get:PathSensitive(PathSensitivity.RELATIVE)
   abstract val snapshotsPath: DirectoryProperty
 
+  @get:Input @get:Optional abstract val vcsHeadSha: Property<String>
+  @get:Input @get:Optional abstract val vcsBaseSha: Property<String>
+  @get:Input @get:Optional abstract val vcsProvider: Property<String>
+  @get:Input @get:Optional abstract val vcsHeadRepoName: Property<String>
+  @get:Input @get:Optional abstract val vcsBaseRepoName: Property<String>
+  @get:Input @get:Optional abstract val vcsHeadRef: Property<String>
+  @get:Input @get:Optional abstract val vcsBaseRef: Property<String>
+  @get:Input @get:Optional abstract val vcsPrNumber: Property<Int>
+
   override fun getArguments(args: MutableList<String>) {
     args.add("build")
     args.add("snapshots")
     args.add("--app-id")
     args.add(appId.get())
+
+    vcsHeadSha.orNull?.let { args.addAll(listOf("--head-sha", it)) }
+    vcsBaseSha.orNull?.let { args.addAll(listOf("--base-sha", it)) }
+    vcsProvider.orNull?.let { args.addAll(listOf("--vcs-provider", it)) }
+    vcsHeadRepoName.orNull?.let { args.addAll(listOf("--head-repo-name", it)) }
+    vcsBaseRepoName.orNull?.let { args.addAll(listOf("--base-repo-name", it)) }
+    vcsHeadRef.orNull?.let { args.addAll(listOf("--head-ref", it)) }
+    vcsBaseRef.orNull?.let { args.addAll(listOf("--base-ref", it)) }
+    vcsPrNumber.orNull?.let { args.addAll(listOf("--pr-number", it.toString())) }
+
     args.add(snapshotsPath.get().asFile.absolutePath)
   }
 
@@ -65,6 +85,14 @@ abstract class SentryUploadSnapshotsTask : SentryCliExecTask() {
         task.sentryUrl.set(extension.url)
         task.appId.set(extension.snapshots.appId)
         task.snapshotsPath.set(extension.snapshots.path)
+        task.vcsHeadSha.set(extension.vcsInfo.headSha)
+        task.vcsBaseSha.set(extension.vcsInfo.baseSha)
+        task.vcsProvider.set(extension.vcsInfo.vcsProvider)
+        task.vcsHeadRepoName.set(extension.vcsInfo.headRepoName)
+        task.vcsBaseRepoName.set(extension.vcsInfo.baseRepoName)
+        task.vcsHeadRef.set(extension.vcsInfo.headRef)
+        task.vcsBaseRef.set(extension.vcsInfo.baseRef)
+        task.vcsPrNumber.set(extension.vcsInfo.prNumber)
         task.asSentryCliExec()
       }
     }